requirement.txt
Django==2.2.5
djangorestframework==3.11.0
mysqlclient==1.4.6
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.2
sqlparse==0.3.0
